client-api: Move get_capabilities into discovery
