+++ crates/ruma-client-api/src/discovery.rs
@@ -1,4 +1,5 @@
-//! Endpoints that cannot change with new versions of the Matrix specification.
+//! Server discovery endpoints.
 
 pub mod discover_homeserver;
+pub mod get_capabilities;
 pub mod get_supported_versions;
