+++ crates/ruma-client-api/src/discovery/get_capabilities/iter.rs
